pr-review: SortedSerializedBase -> models/base.py
